requirements.txt
Flask
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-telegram-bot
pymongo
gunicorn
python-dotenv
nest-asyncio